[SCMenu] 컴파일 경고 제거 — size_t 절단 캐스팅
서브버튼 역방향 루프 인덱스·get_menu_count 반환 size_t→int 명시.

+++ CMenu/CSCMenuBar/SCMenu.cpp
@@ -2106,7 +2106,7 @@ void CSCMenu::OnPaint()
 			//sub button: 컨텐츠 좌표는 m_r 에 저장 (hit test용), 그리기는 view 좌표.
 			CRect r_content = m_items[i]->m_r;
 			CRect r_view = view_r;
-			for (int j = m_items[i]->m_buttons.size() - 1; j >= 0; j--)
+			for (int j = (int)(m_items[i]->m_buttons.size() - 1); j >= 0; j--)
 			{
 				int btn_w = m_items[i]->m_buttons[j]->m_r.Width();
 				r_content.right -= 6;
